Simplify function params
Previously, the parameters for function calls were configured using a
custom abstraction cooked up by the AI in initial passes for this
library. However, the API was very limited. Also, for a custom
abstraction like that, I'd prefer to have a strong vision for something
I believe is an improvement over OpenAI's API.

I don't have that, so for now (an maybe indefinitely?) I'm just going to
pass through parameters to OpenAI's API.