feat: add .chat to agent
I'm thinking it might make sense for the duck type of Angents and
Agencies to just be:

1. It responds to .chat
2. It responds to .chat
3. These functions return an array of hashes with role and content keys

If we can do that, it makes the APIs maximally generalizable. For
example, when passing in a list of agents to an agency, those items in
the array could be:

1. The class of an agent
2. An instance of an agent
3. The class of an agency
4. An instance of an agency

This should make constructing agencies much easier.